AlphaTestbed now instantiates metrics and has a separate metric instance for each service
The metrics are now instantiated in the AlphaTestbed class. This meant that I also had to move the metrics parameters to the constructor of the AlphaTestbed. 

The AlphaTestbed now stores the instances of metrics in a HashMap. In this map, keys represent services and values represent metric instances. There are two different maps in the AlphaTestbed class. One map stores ranking metrics and the other stores the utility metrics. This is required because a metric instance may be stateful. In such cases, using the same metric instance across different services would yield incorrect results. For this reason, each service has its own metric instance.

Note that the //current evaluation results// are stored in a map, where keys represent a //combination of a service and a metric// (computed by xoring the hashes of the metric class and the service type) and values represent the current value of the metric. This is required so that different subscribers are able to subscribe to the same combination of the metric/service.

This now closes #6.